refactor(debug): promote Notifications to its own Developer entry
