+++ swift/TigerDuck/Debug/DebugSettingsViewModel.swift
@@ -1,7 +1,6 @@
 #if DEBUG
 import Foundation
 import Observation
-import UserNotifications
 
 @MainActor
 @Observable
@@ -10,7 +9,6 @@ final class DebugSettingsViewModel {
     var draftInstant: Date
     var frozen: Bool
     private(set) var effectiveNow: Date
-    private(set) var lastSimulatedPushStatus: String?
 
     init() {
         let current = DebugClockController.shared.currentOverride()
@@ -62,47 +60,5 @@ final class DebugSettingsViewModel {
         DebugClockController.shared.setOverride(override)
     }
 
-    /// Fires a local notification ~3 s from real-now to simulate what a
-    /// backend APNs push would look like. Backend pushes can't honor the
-    /// debug clock (the server has no idea time is faked), so this exists
-    /// purely so QA can visually confirm a "push arrived" while the LA /
-    /// widget pipeline is running under a fake-clock override.
-    func sendSimulatedPushNow() async {
-        let center = UNUserNotificationCenter.current()
-        let settings = await center.notificationSettings()
-        switch settings.authorizationStatus {
-        case .notDetermined:
-            let granted = (try? await center.requestAuthorization(options: [.alert, .sound])) ?? false
-            guard granted else {
-                lastSimulatedPushStatus = "Authorization denied"
-                return
-            }
-        case .denied:
-            lastSimulatedPushStatus = "Authorization denied — enable in Settings"
-            return
-        case .authorized, .provisional, .ephemeral:
-            break
-        @unknown default:
-            break
-        }
-
-        let content = UNMutableNotificationContent()
-        content.title = "Simulated push"
-        content.body = "Fake push fired at \(AppClock.now().formatted(date: .omitted, time: .standard)) (app clock)"
-        content.sound = .default
-
-        let trigger = UNTimeIntervalNotificationTrigger(timeInterval: 3, repeats: false)
-        let request = UNNotificationRequest(
-            identifier: "debug-simulated-push-\(UUID().uuidString)",
-            content: content,
-            trigger: trigger
-        )
-        do {
-            try await center.add(request)
-            lastSimulatedPushStatus = "Scheduled — arrives in ~3 s (real time)"
-        } catch {
-            lastSimulatedPushStatus = "Failed: \(error.localizedDescription)"
-        }
-    }
 }
 #endif
